services/impression-agg/requirements.txt
pyspark==2.4.4
mysql-connector==2.2.9